src/bot.py
"""
Telegram bot handler
Manages registration and login confirmation via Telegram
"""
from telegram import Update
from telegram.ext import Application, CommandHandler, ContextTypes
from src.config import settings
from src.services.auth_service import AuthService
from src.services.token_service import TokenService

class TeleLoginBot:
    def __init__(self):
        self.app = Application.builder().token(settings.BOT_TOKEN).build()
        self.auth_service = AuthService()
        self.token_service = TokenService()
        
    async def start_command(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /start command with registration token"""
        if context.args:
            token = context.args[0]
            telegram_id = update.effective_user.id
            # Process registration token
            # TODO: Implement token verification and user linking
            await update.message.reply_text("Registration successful!")
        else:
            await update.message.reply_text("Welcome to TeleLogin!")
    
    async def confirm_login(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle login confirmation callback"""
        # TODO: Implement login confirmation logic
        pass
    
    def run(self):
        """Start the bot"""
        self.app.add_handler(CommandHandler("start", self.start_command))
        self.app.run_polling()

if __name__ == "__main__":
    bot = TeleLoginBot()
    bot.run()
